requirements.txt
# Web Framework
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0

# Web Scraping
camoufox==0.2.0
html2text==2024.2.26

# Type Safety & Validation
pydantic==2.5.0
typing-extensions==4.8.0

# Utilities
python-decouple==3.8
marshmallow==3.20.2

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
gevent==23.9.1